requirements.txt
argon2-cffi==23.1.0
Faker==25.8.0
Flask==3.0.3
matplotlib==3.9.0
networkx==3.3
pymongo==4.7.3
pynauty==2.8.8.1
python-dotenv==1.0.1
PyQt5==5.15.10
requests==2.32.3
